attention_gan_pretrained/requirements.txt
torch
torchvision
numpy
opencv-python-headless
matplotlib
tqdm
albumentations
PyYAML
scikit-learn
streamlit
